refactor: change name to match standard. Change Save Input to Retrieve Past Inputs
